chore: move TokenCore type to pool.types
